+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug fixes 
 1. [#132](https://github.com/influxdata/influxdb-client-go/pull/132) Properly handle errors instead of panics
+1. [#134](https://github.com/influxdata/influxdb-client-go/pull/134) FluxQueryResult: support reordering of annotations
 
 ## 1.2.0 [2020-05-15]
 ### Breaking Changes