+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.3.0 [in progress]
+## 1.3.0 [2020-06-19]
 ### Features
 1. [#131](https://github.com/influxdata/influxdb-client-go/pull/131) Labels API
 1. [#136](https://github.com/influxdata/influxdb-client-go/pull/136) Possibility to specify default tags